pulldown_cmark_to_cmark

Function cmark_resume_with_options

pub fn cmark_resume_with_options<'a, I, E, F>(
    events: I,
    formatter: F,
    state: Option<State<'a>>,
    options: Options<'_>,
) -> Result<State<'a>, Error>
where
    I: Iterator<Item = E>,
    E: Borrow<Event<'a>>,
    F: Write,

Serialize a stream of pulldown-cmark-Events into a string-backed buffer.

events

An iterator over Events, for example as returned by the Parser

formatter

A format writer, can be a String.

state

The optional initial state of the serialization.

options

Customize the appearance of the serialization. All otherwise magic values are contained here.

Returns the State of the serialization on success. You can use it as initial state in the next call if you are halting event serialization.

Errors if the underlying buffer fails (which is unlikely) or if the Event stream cannot ever be produced by deserializing valid Markdown. Each failure mode corresponds to one of Error’s variants.